% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_surv_survivalsvm.R
\name{LearnerSurvSurvivalsvm}
\alias{LearnerSurvSurvivalsvm}
\title{R6 Class to construct a Support Vector Machine survival learner}
\description{
The `LearnerSurvSurvivalsvm` class is the interface to the `survivalsvm`
  R package for use with the `mlexperiments` package.
}
\details{
Optimization metric: C-index
Can be used with
* [mlexperiments::MLTuneParameters]
* [mlexperiments::MLCrossValidation]
* [mlexperiments::MLNestedCVs]
}
\examples{
# survival analysis

dataset <- survival::colon |>
  data.table::as.data.table() |>
  na.omit()
dataset <- dataset[get("etype") == 2, ]

seed <- 123
surv_cols <- c("status", "time", "rx")

feature_cols <- colnames(dataset)[3:(ncol(dataset) - 1)]

param_list_survivalsvm <- expand.grid(
  sample.fraction = seq(0.6, 1, .2),
  min.node.size = seq(1, 5, 4),
  mtry = seq(2, 6, 2),
  num.trees = c(5L, 10L),
  max.depth = seq(1, 5, 4)
)

ncores <- 2L

split_vector <- splitTools::multi_strata(
  df = dataset[, .SD, .SDcols = surv_cols],
  strategy = "kmeans",
  k = 4
)

train_x <- model.matrix(
  ~ -1 + .,
  dataset[, .SD, .SDcols = setdiff(feature_cols, surv_cols[1:2])]
)
train_y <- survival::Surv(
  event = (dataset[, get("status")] |>
             as.character() |>
             as.integer()),
  time = dataset[, get("time")],
  type = "right"
)

fold_list <- splitTools::create_folds(
  y = split_vector,
  k = 3,
  type = "stratified",
  seed = seed
)

surv_survivalsvm_optimizer <- mlexperiments::MLCrossValidation$new(
  learner = mllrnrs::LearnerSurvSurvivalsvm$new(),
  fold_list = fold_list,
  ncores = ncores,
  seed = seed
)
surv_survivalsvm_optimizer$learner_args <- as.list(
  data.table::data.table(param_list_survivalsvm[1, ],
                         stringsAsFactors = FALSE)
)
surv_survivalsvm_optimizer$performance_metric <- c_index

# set data
surv_survivalsvm_optimizer$set_data(
  x = train_x,
  y = train_y
)

surv_survivalsvm_optimizer$execute()


## ------------------------------------------------
## Method `LearnerSurvSurvivalsvm$new`
## ------------------------------------------------

LearnerSurvSurvivalsvm$new()

}
\seealso{
[survivalsvm::survivalsvm()]
}
\section{Super class}{
\code{\link[mlexperiments:MLLearnerBase]{mlexperiments::MLLearnerBase}} -> \code{LearnerSurvSurvivalsvm}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvSurvivalsvm-new}{\code{LearnerSurvSurvivalsvm$new()}}
\item \href{#method-LearnerSurvSurvivalsvm-clone}{\code{LearnerSurvSurvivalsvm$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="bayesian_scoring_function"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-bayesian_scoring_function'><code>mlexperiments::MLLearnerBase$bayesian_scoring_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="cross_validation"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-cross_validation'><code>mlexperiments::MLLearnerBase$cross_validation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="fit"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-fit'><code>mlexperiments::MLLearnerBase$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="predict"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-predict'><code>mlexperiments::MLLearnerBase$predict()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvSurvivalsvm-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvSurvivalsvm-new}{}}}
\subsection{Method \code{new()}}{
Create a new `LearnerSurvSurvivalsvm` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvSurvivalsvm$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `LearnerSurvSurvivalsvm` R6 object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{LearnerSurvSurvivalsvm$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvSurvivalsvm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvSurvivalsvm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvSurvivalsvm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
